+++ Client Source Code/WarFare/UITradeSellBBS.cpp
@@ -377,8 +377,8 @@ void CUITradeSellBBS::MsgSend_Register()
 	iPrice			= CGameProcedure::s_pProcMain->m_pUITradeBBSEdit->GetPrice();
 
 	sLen = 15;
-	sLen += szTitle.size();
-	sLen += szExplanation.size();
+	sLen += (short)szTitle.size();
+	sLen += (short)szExplanation.size();
 
 	BYTE* byBuff = new BYTE[sLen];
 	int iOffset=0;
@@ -386,9 +386,9 @@ void CUITradeSellBBS::MsgSend_Register()
 	CAPISocket::MP_AddByte(byBuff, iOffset, N3_TRADE_BBS);	
 	CAPISocket::MP_AddByte(byBuff, iOffset, N3_SP_TYPE_REGISTER);
 	CAPISocket::MP_AddByte(byBuff, iOffset, m_byBBSKind);
-	CAPISocket::MP_AddShort(byBuff, iOffset, szTitle.size());
+	CAPISocket::MP_AddShort(byBuff, iOffset, (short)szTitle.size());
 	CAPISocket::MP_AddString(byBuff, iOffset, szTitle);
-	CAPISocket::MP_AddShort(byBuff, iOffset, szExplanation.size());
+	CAPISocket::MP_AddShort(byBuff, iOffset, (short)szExplanation.size());
 	CAPISocket::MP_AddString(byBuff, iOffset, szExplanation);
 	CAPISocket::MP_AddDword(byBuff, iOffset, iPrice);
 	CGameProcedure::s_pSocket->Send(byBuff, iOffset);
@@ -682,10 +682,10 @@ void CUITradeSellBBS::RenderSelectContent()
 		return;
 
 	__VertexTransformedColor vLines[5];
-	vLines[0].Set(rc.left, rc.top, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
-	vLines[1].Set(rc.right, rc.top, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
-	vLines[2].Set(rc.right, rc.bottom, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
-	vLines[3].Set(rc.left, rc.bottom, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
+	vLines[0].Set((float)rc.left, (float)rc.top, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
+	vLines[1].Set((float)rc.right, (float)rc.top, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
+	vLines[2].Set((float)rc.right, (float)rc.bottom, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
+	vLines[3].Set((float)rc.left, (float)rc.bottom, UI_DEFAULT_Z, UI_DEFAULT_RHW, 0xff00ff00);
 	vLines[4] = vLines[0];
 
 	DWORD dwZ, dwFog, dwAlpha, dwCOP, dwCA1, dwSrcBlend, dwDestBlend, dwVertexShader, dwAOP, dwAA1;
@@ -733,7 +733,7 @@ bool CUITradeSellBBS::SelectedString(CN3UIBase* pSender, int& iID)
 		if(pSender == m_pText[i])
 		{
 			iIndex = i % TRADE_BBS_MAX_LINE;
-			if( iIndex >= m_Datas.size() )
+			if( iIndex >= (int)m_Datas.size() )
 				return false;
 
 			iID = iIndex;
